program structure + inverse kinematics
added program structure and robust user interface,
included Michael's inverse kinematics
organization and code structure
