+++ source/operation.h
@@ -1,0 +1,12 @@
+#ifndef OPERATION_H_INCLUDED
+#define OPERATION_H_INCLUDED
+
+Registerspace xyz_set_joints(Registerspace d);
+int set_speed();
+Registerspace manual_set_joints(Registerspace d);
+void forward_bot1();
+void inverse_bot();
+void main_menu();
+
+
+#endif // OPERATION_H_INCLUDED
